[ALPS03216271] ION: fix coverity scanned code defect
[Detail] fix improper use of nagative value

[Solution]
fix ion_mm_heap_get_phys and ion address not equal

MTK-Commit-Id: 4907a0c4c98dca532b7fc33688d5ce947772a8fe

CR-Id: ALPS03216271
Feature: others

+++ drivers/staging/android/ion/ion.c
@@ -595,7 +595,7 @@ int ion_phys(struct ion_client *client, struct ion_handle *handle,
 	ret = buffer->heap->ops->phys(buffer->heap, buffer, addr, len);
 
 	/*avoid camelcase, will modify in a letter*/
-	mmprofile_log_ex(ion_mmp_events[PROFILE_GET_PHYS], MMPROFILE_FLAG_END, buffer->size, *addr);
+	mmprofile_log_ex(ion_mmp_events[PROFILE_GET_PHYS], MMPROFILE_FLAG_END, buffer->size, (unsigned long)*addr);
 
 	return ret;
 }
